requirements.txt
pandas
numpy
scikit-learn
groq
python-dotenv
